+++ src/components/transaction/TransactionList/transaction-list.js
@@ -32,6 +32,11 @@ const visibleColumns = [
   TRANSACTION_FIELDS.STATUS,
 ];
 
+const nonInteractiveHeaders = [
+  TRANSACTION_FIELDS.ITEM_TITLE,
+  TRANSACTION_FIELDS.PATRON_NAME,
+];
+
 const columnMapping = {
   [TRANSACTION_FIELDS.TIME]: <FormattedMessage id="ui-inn-reach.transaction.field.time" />,
   [TRANSACTION_FIELDS.TYPE]: <FormattedMessage id="ui-inn-reach.transaction.field.type" />,
@@ -132,6 +137,7 @@ const TransactionListView = ({
       id="transactions-list"
       isLoading={isLoading}
       isShowAddNew={false}
+      nonInteractiveHeaders={nonInteractiveHeaders}
       resetData={resetData}
       pagingSlipsArr={pagingSlipsArr}
       pagingSlipTemplatesMap={pagingSlipTemplatesMap}